Updated means of activating/Deactivating showers
