--- Content.Server/_Den/Shower/ShowerSystem.cs
@@ -1,9 +1,0 @@
-/// Made by Sleepy Scarecrow for Denstation
-using Content.Shared._Den.Showers;
-
-namespace Content.Server._Den.Shower;
-
-public sealed class ShowerSystem : SharedShowerSystem
-{
-
-}
