--- Content.Shared/_Den/ShowerComponent.cs
@@ -1,43 +1,0 @@
-/// Made by Sleepy Scarecrow for Denstation
-using Robust.Shared.Audio;
-using Robust.Shared.GameStates;
-using Robust.Shared.Serialization;
-
-namespace Content.Shared._Den.Showers
-{
-    /// <summary>
-    /// showers that can be enabled
-    /// </summary>
-    [RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-    public sealed partial class ShowerComponent : Component
-    {
-        /// <summary>
-        /// Toggles shower.
-        /// </summary>
-        [DataField, AutoNetworkedField]
-        public bool ToggleShower;
-
-        [DataField("enableShowerSound")]
-        public SoundSpecifier EnableShowerSound = new SoundPathSpecifier("/Audio/_DEN/Ambience/Objects/shower_enable.ogg");
-
-        public EntityUid? PlayingStream;
-
-        [DataField("loopingSound")]
-        public SoundSpecifier LoopingSound = new SoundPathSpecifier("/Audio/_DEN/Ambience/Objects/shower_running.ogg");
-
-    }
-
-
-    [Serializable, NetSerializable]
-    public enum ShowerVisuals : byte
-    {
-        ShowerVisualState,
-    }
-
-    [Serializable, NetSerializable]
-    public enum ShowerVisualState : byte
-    {
-        Off,
-        On
-    }
-}
