+++ Resources/Audio/_Triad/Ambience/Objects/shower_enable.ogg
