+++ Resources/Prototypes/_Triad/Entities/Structures/shower.yml
@@ -8,7 +8,7 @@
     - Wallmount
   components:
     - type: Sprite
-      sprite: _Triad/Structure/Furniture/shower.rsi
+      sprite: _DEN/Structures/Furniture/shower.rsi
       drawdepth: WallMountedItems
       layers:
       - state: shower
@@ -21,19 +21,20 @@
     - type: Clickable
     - type: ItemToggle
       soundActivate:
-        path: /Audio/_Triad/Ambience/Objects/shower_enable.ogg
+        path: /Audio/_DEN/Ambience/Objects/shower_enable.ogg
       soundDeactivate:
-        path: /Audio/_Triad/Ambience/Objects/shower_enable.ogg
-        predicted: false
+        path: /Audio/_DEN/Ambience/Objects/shower_enable.ogg
     - type: ItemToggleActiveSound
       activeSound:
-        path: /Audio/_Triad/Ambience/Objects/shower_running.ogg
+        path: /Audio/_DEN/Ambience/Objects/shower_running.ogg
     - type: UseDelay
       delay: 1.0
+    - type: ItemToggleRandomToggle
+      chance: 0.1 # 10% chance to start toggled on
     - type: Physics
       bodyType: Static
       canCollide: false
-    - type: InteractionThOutline
+    - type: InteractionOutline
     - type: Damageable
     - type: Construction
       graph: Shower
