+++ Resources/Prototypes/_Triad/Recipes/Construction/Graphs/furniture/shower.yml
