+++ Resources/Prototypes/_Triad/Recipes/Construction/furniture.yml
@@ -7,7 +7,7 @@
   category: construction-category-furniture
   description: A shower, complete with bluespace water.
   icon:
-    sprite: _DEN/Structures/Furniture/shower.rsi
+    sprite: _Triad/Structure/Furniture/shower.rsi
     state: shower
     objectType: Structure
   placementMode: SnapgridCenter
