+++ Resources/Textures/_Triad/Structure/Furniture/shower.rsi/meta.json
