+++ Resources/Textures/_Triad/Structure/Furniture/shower.rsi/shower.png
