+++ Resources/Textures/_Triad/Structure/Furniture/shower.rsi/shower_on.png
